configuration, types: Add support for sweeps
Sweeps are handled by one of three handlers, ``autofreq``, ``range`` and
``autoparam``. ``range`` allows the user to manually specify values for
a parameter in the agenda; ``autofreq`` allows cpu frequencies to be
swept through, and autoparam allows other parameters to be swept
through.

autoparam requries the parameter to have ``allowed_values`` specified,
and the sweep specification must supply the plugin name and parameter
name as:
``
	sweep(autoparam):
		param: <param name>
		plugin: <plugin name>
``

autofreq can be used with an empty value:
``
	sweep(autofreq):
``
to sweep through all values of ``frequency`` by default, although if
``plugin`` is specified, other parameters within cpufreq can be swept
through:
``
	sweep(autofreq):
		param: cpu0_frequency
``

For either of the above 'automatic' sweeps, a minimum and/or maximum can
be specified to limit the values generated in the sweep:
``
	sweep(autofreq):
		param: cpu0_frequency
		min: 1000000
		max: 1700000
``

``range`` sweeps support two specification syntaxes: one for manual
specification of a list, and another for a range of values. These sweeps
only accept the plugin name as a parameter:
``
	sweep(range):
		threads: [1, 3, 5, 7]
``
or
``
	sweep(range):
		threads: 1-8,2
		# start-stop[,step]
``
These both produce the same result. step is optional and defaults to 1.

Any sweep specified in a workload entry is equivalent to manually
specifying individual workload entries, each with one of the possible
values of the sweep.

Sweeps specified in sections will create the equivalent of a section
group, each section in that group with one of the sweep values. When the
group name is specified, the 'expanded' section entry will maintain that
group name, otherwise will generate its own.

If multiple sweeps are specified in one entry, then this will be
expanded to one entry for every combination of values of each sweep.