instruments/trace_cmd: Support setting trace-cmd mode
The doc comment for the instrument mentions the option to select the
trace_cmd_mode ('start' or 'record'). As far as I can tell this was
never implemented. This commit adds the support for passing
trace_cmd_mode to the FtraceCollector in devlib.

While at it, it also does some slight cleanup aronud the instrument,
e.g. in the teardown() implementation which should be folded
inside the collector instead of reimplemented here.

Signed-off-by: Kajetan Puchalski <kajetan.puchalski@arm.com>