workloads: Update jetnews wa tests
The jetnews wa workload runs a single test at the moment, but there are
more tests available.

This patch enables two additional basic tests and renames the wa workload
name to "jetnews", in preparation to potential future updates where we will
be able to select the kind of test we want to run.