[workloads/jetnews] Bump the minSdkVersion to 28 and targetSdkVersion to 34
API level 27 (Android 8.1/Oreo) doesn't report a frame metric called
"deadline missed", but the JankTest library expects to be able to find
data for that metric. If it is not found, an assertion is reported and
the run fails.

Given Android Oreo is reasonably old at this point, set the minimum SDK
version to 28 so we can gather all the metrics we need.

Also bump the targetSdkVersion to make it clear what version was used
for putting this code together.